requirements.txt
PyMuPDF
chromadb
openai
rank_bm25
pydantic
fastapi
uvicorn
python-dotenv
python-multipart
numpy
